Rename extension.ts to UI and make it clear it's for UI Extensions
Right now this is still used by theme extensions.  I'll change that in the next commit
