+++ packages/app/src/cli/models/app/app.test-data.ts
@@ -1,6 +1,6 @@
 import {App, AppInterface} from './app.js'
 import {FunctionExtension, ThemeExtension, UIExtension} from './extensions.js'
-import {ExtensionInstance, specForType} from '../extensions/extensions.js'
+import {UIExtensionInstance, specForType} from '../extensions/ui.js'
 import {FunctionInstance, functionSpecForType} from '../extensions/functions.js'
 
 export function testApp(app: Partial<AppInterface> = {}): AppInterface {
@@ -46,7 +46,7 @@ export async function testUIExtension(uiExtension: Partial<UIExtension> = {}): P
 
   const specification = await specForType(configuration.type)
 
-  const extension = new ExtensionInstance({
+  const extension = new UIExtensionInstance({
     configuration,
     configurationPath,
     entryPath: entrySourceFilePath,
@@ -66,7 +66,7 @@ export async function testThemeExtensions(): Promise<ThemeExtension> {
   }
 
   const specification = await specForType(configuration.type)
-  return new ExtensionInstance({
+  return new UIExtensionInstance({
     configuration,
     configurationPath: '',
     entryPath: '',
