+++ packages/app/src/cli/models/extensions/extension-specifications/post_purchase_extension.ts
@@ -1,10 +1,10 @@
-import {createExtensionSpec} from '../extensions.js'
+import {createUIExtensionSpec} from '../ui.js'
 import {BaseExtensionSchema} from '../schemas.js'
 import {output} from '@shopify/cli-kit'
 
 const dependency = {name: '@shopify/post-purchase-ui-extensions-react', version: '^0.13.2'}
 
-const spec = createExtensionSpec({
+const spec = createUIExtensionSpec({
   identifier: 'checkout_post_purchase',
   externalIdentifier: 'post_purchase_ui',
   externalName: 'Post-purchase UI',
