+++ packages/app/src/cli/models/extensions/extension-specifications/product_subscription.ts
@@ -1,11 +1,11 @@
 import {getDependencyVersion} from '../../app/app.js'
-import {createExtensionSpec} from '../extensions.js'
+import {createUIExtensionSpec} from '../ui.js'
 import {BaseExtensionSchema} from '../schemas.js'
 import {error} from '@shopify/cli-kit'
 
 const dependency = {name: '@shopify/admin-ui-extensions-react', version: '^1.0.1'}
 
-const spec = createExtensionSpec({
+const spec = createUIExtensionSpec({
   identifier: 'product_subscription',
   externalIdentifier: 'subscription_ui',
   externalName: 'Subscription UI',
