+++ packages/app/src/cli/models/extensions/extension-specifications/web_pixel_extension.ts
@@ -1,4 +1,4 @@
-import {createExtensionSpec} from '../extensions.js'
+import {createUIExtensionSpec} from '../ui.js'
 import {BaseExtensionSchema} from '../schemas.js'
 import {error, schema} from '@shopify/cli-kit'
 
@@ -11,7 +11,7 @@ const WebPixelSchema = BaseExtensionSchema.extend({
   settings: schema.define.any(),
 })
 
-const spec = createExtensionSpec({
+const spec = createUIExtensionSpec({
   identifier: 'web_pixel_extension',
   externalIdentifier: 'web_pixel',
   externalName: 'Web pixel',
