+++ packages/app/src/cli/models/extensions/ui.ts
@@ -10,7 +10,7 @@ export type BaseConfigContents = schema.define.infer<typeof BaseExtensionSchema>
 /**
  * Extension specification with all the needed properties and methods to load an extension.
  */
-export interface ExtensionSpec<TConfiguration extends BaseConfigContents = BaseConfigContents>
+export interface UIExtensionSpec<TConfiguration extends BaseConfigContents = BaseConfigContents>
   extends ExtensionIdentifier {
   identifier: string
   externalIdentifier: string
@@ -49,7 +49,7 @@ export interface ExtensionSpec<TConfiguration extends BaseConfigContents = BaseC
  *
  * This class holds the public interface to interact with extensions
  */
-export class ExtensionInstance<TConfiguration extends BaseConfigContents = BaseConfigContents>
+export class UIExtensionInstance<TConfiguration extends BaseConfigContents = BaseConfigContents>
   implements UIExtension<TConfiguration>, ThemeExtension<TConfiguration>
 {
   entrySourceFilePath: string
@@ -61,7 +61,7 @@ export class ExtensionInstance<TConfiguration extends BaseConfigContents = BaseC
   configuration: TConfiguration
   configurationPath: string
 
-  private specification: ExtensionSpec
+  private specification: UIExtensionSpec
   private remoteSpecification?: api.graphql.RemoteSpecification
 
   get graphQLType() {
@@ -101,7 +101,7 @@ export class ExtensionInstance<TConfiguration extends BaseConfigContents = BaseC
     configurationPath: string
     entryPath: string
     directory: string
-    specification: ExtensionSpec
+    specification: UIExtensionSpec
     remoteSpecification?: api.graphql.RemoteSpecification
   }) {
     this.configuration = options.configuration
@@ -169,7 +169,7 @@ export class ExtensionInstance<TConfiguration extends BaseConfigContents = BaseC
 /**
  * Find the registered spececification for a given extension type
  */
-export async function specForType(type: string): Promise<ExtensionSpec | undefined> {
+export async function specForType(type: string): Promise<UIExtensionSpec | undefined> {
   const allSpecs = await allExtensionSpecifications()
   return allSpecs.find((spec) => spec.identifier === type || spec.externalIdentifier === type)
 }
@@ -179,7 +179,7 @@ function remoteSpecForType(type: string): api.graphql.RemoteSpecification | unde
   return undefined
 }
 
-export function createExtensionSpec<TConfiguration extends BaseConfigContents = BaseConfigContents>(spec: {
+export function createUIExtensionSpec<TConfiguration extends BaseConfigContents = BaseConfigContents>(spec: {
   identifier: string
   externalIdentifier: string
   partnersWebIdentifier: string
@@ -203,7 +203,7 @@ export function createExtensionSpec<TConfiguration extends BaseConfigContents = 
   ) => output.TokenizedString | undefined
   shouldFetchCartUrl?(config: TConfiguration): boolean
   hasExtensionPointTarget?(config: TConfiguration, target: string): boolean
-}): ExtensionSpec<TConfiguration> {
+}): UIExtensionSpec<TConfiguration> {
   const defaults = {
     showInCLIHelp: true,
     singleEntryPath: true,
