+++ packages/app/src/cli/plugins/extension.ts
@@ -1,4 +1,4 @@
-import {ExtensionSpec} from '../models/extensions/extensions.js'
+import {UIExtensionSpec} from '../models/extensions/ui.js'
 import {FunctionSpec} from '../models/extensions/functions.js'
 import {plugins} from '@shopify/cli-kit'
 import {Config} from '@oclif/core'
@@ -13,7 +13,7 @@ interface HookReturnPerExtensionPlugin extends plugins.HookReturnsPerPlugin {
   extension_specs: {
     options: {[key: string]: never}
     pluginReturns: {
-      [pluginName: string]: ExtensionSpec[]
+      [pluginName: string]: UIExtensionSpec[]
     }
   }
   function_specs: {
@@ -27,7 +27,7 @@ interface HookReturnPerExtensionPlugin extends plugins.HookReturnsPerPlugin {
 export type ExtensionSpecsFunction = plugins.FanoutHookFunction<'extension_specs', ''>
 export type FunctionSpecsFunction = plugins.FanoutHookFunction<'function_specs', ''>
 
-export const defineExtensionSpecs = (input: ExtensionSpec): ExtensionSpecsFunction => {
+export const defineExtensionSpecs = (input: UIExtensionSpec): ExtensionSpecsFunction => {
   return async () => input
 }
 
@@ -35,7 +35,7 @@ export const defineFunctionSpecs = (input: FunctionSpec): FunctionSpecsFunction 
   return async () => input
 }
 
-export async function getListOfExtensionSpecs(config: Config): Promise<ExtensionSpec[]> {
+export async function getListOfExtensionSpecs(config: Config): Promise<UIExtensionSpec[]> {
   const hooks = await plugins.fanoutHooks<HookReturnPerExtensionPlugin, 'extension_specs'>(
     config,
     'extension_specs',
